Fixed all warnings for Engine and Editor
+ Updated various vector types to use int
+ Fixed API macro and vector/template exporting
+ Various conversion loss warning fixes
+ Added disable warning for a specific issue with std libraries to fix/improve later
+ Added glew static macro to premake
+ Added hpp files to premake

+++ SkyEngine/Graphics/GLFW/GLAPI.cpp
@@ -173,7 +173,7 @@ bool IGLAPI::CreateShader(uint32_t& ShaderID, GLenum shaderType, std::string sou
 
 	ShaderID = glCreateShader(shaderType);
 	const char* shader_code_ptr = source.c_str();
-	const int shader_code_size = source.size();
+	const int shader_code_size = static_cast<int>(source.size());
 
 	glShaderSource(ShaderID, 1, &shader_code_ptr, &shader_code_size);
 	glCompileShader(ShaderID);
